# SQL performance improvements