… UPGRADING.md
… src/Security/CurrentUser.php
… src/Timesheet/UserDateTimeFactory.php
… tests/Mocks/Security/CurrentUserFactory.php
… tests/Repository/Loader/ActivityLoaderTest.php