## Description

This PR will bring a massive performance boost to every installation for: tags, users, team-timesheets, reports and in general every page load. Small installations might not see all of these improvements, the most benefit will be visible in installations with a lot of records.

**Test setup for comparison:**
- Timesheets: > 980000
- Tags: > 53000
- Users: > 400 (only 200 had records)
- Activities: > 29000  (only 350 were used in timesheets)
- Projects: > 4360 (only 50 were used in timesheets)
- Customers: > 350 (only 5 were used in timesheets)

**Loading times before (master):**
- My times (as teamlead, 4100 records): 22 queries / 235ms 
- Report "all users one month" (as teamlead with ~50 users): 72 queries / 1000ms
- Report "monthly view one user" (as teamlead, full month 55h): 14 queries / 175 ms
- Team timesheets (as Super-Admin): 17 queries / 615ms
- Team timesheets (as Teamlead, 397500 records): 23 queries / 47500ms (invalid pagination and results)
- User administration (as Super-Admin): 12 queries / 40ms
- Dashboard (as Super-Admin): 267 queries / 750 ms
- Dashboard (as Teamlead): 43 queries / 870ms
- Tag listing: 7 queries / 7300ms
- Customer listing (as teamlead): 11 queries / 200 ms
- Project listing (as teamlead): 14 queries / 215 ms
- Activity listing (as teamlead): 17 queries / 310 ms

**Loading times afterwards (this branch):**
- My times (as teamlead, 4100 records): 25 queries / 70ms
- Report "all users one month" (as teamlead with ~50 users): 75 queries / 900ms
- Report "monthly view one user" (as teamlead, full month 55h): 17 queries / 20ms
- Team timesheets (as Super-Admin): 20 queries / 160 ms
- Team timesheets (as Teamlead, 397500 records): 22 queries / 650ms
- User administration (as Super-Admin): 15 queries / 10ms
- Dashboard (as Super-Admin): 270 queries (member of 60 teams & projects) / 750ms
- Dashboard (as Teamlead): 46 queries / 720ms
- Tag listing: 10 queries / 10ms
- Customer listing (as teamlead): 14 queries / 10 ms
- Project listing (as teamlead): 17 queries / 20 ms
- Activity listing (as teamlead): 20 queries / 25 ms

Fixes #1948 
Fixes #2012

### Database tuning
Please note, that with such a large setup you might want to tune your database, to prevent excessive read from mass storage.

There are many tips for a better MySQL performance out there. 
A really simple one with immediately positive effects is the `innodb_buffer_pool_size` setting, which is by default set to 128MB.

You can test the current value with:
```sql 
SELECT @@innodb_buffer_pool_size;
-- or get in MB
SELECT @@innodb_buffer_pool_size/1024/1024;
```

Now to change that to 1GB you would execute:
```sql 
SET GLOBAL innodb_buffer_pool_size=1342177280;
```
And if that doesn't work, you have to adjust your my.ini (please refer to the database documentation, this is not part of Kimai).


## Types of changes
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist
- [x] I verified that my code applies to the guidelines (`composer code-check`)
- [ ] I updated the documentation (see [here](https://github.com/kimai/www.kimai.org/tree/master/_documentation))
- [x] I agree that this code is used in Kimai and will be published under the [MIT license](https://github.com/kevinpapst/kimai2/blob/master/LICENSE)